Handle service and URI restoration in plugin
