+++ plugins/okta-auth.client.ts
@@ -5,7 +5,7 @@ allowing components to interact with a shared OktaAuth instance.
 
 import { Plugin } from '@nuxt/types'
 
-import { OktaAuth } from '@okta/okta-auth-js'
+import { OktaAuth, toRelativeUrl } from '@okta/okta-auth-js'
 
 declare module 'vue/types/vue' {
   interface Vue {
@@ -26,6 +26,24 @@ const oktaPlugin: Plugin = (_ctx, inject) => {
 
   // Create OktaAuth instance
   const oktaAuth = new OktaAuth(configOptions)
+
+  // Start the auth background service
+  console.debug('Starting OktaAuth service...')
+  oktaAuth.start()
+
+  // Look for a router instance to handle URI restoration
+  const router = _ctx.app.router
+  if (!oktaAuth.options.restoreOriginalUri) {
+    // eslint-disable-next-line require-await
+    oktaAuth.options.restoreOriginalUri = async (_: OktaAuth, originalUri: string) => {
+      console.log('restoring URI')
+      // If a router is available, provide a default implementation
+      if (router && originalUri) {
+        const path = toRelativeUrl(originalUri, window.location.origin)
+        router.replace({ path })
+      }
+    }
+  }
 
   // Inject oktaAuth into Vue
   inject('okta', oktaAuth)
